🧠 refactor: api 경로 home -> 각 엔티티에 맞는 위치로 수정
issue: #40
